# openapi-schema.yml
openapi: 3.0.3
info:
  title: WorkBridge API
  version: 1.0.0
  description: AI-powered location-based temporary job platform for rural workers.
paths:
...
  /api/v1/auth/admin-login/:
    post:
      operationId: api_v1_auth_admin_login_create
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Login'
          description: ''
...
  /api/v1/auth/demo-tokens/:
    get:
      operationId: api_v1_auth_demo_tokens_retrieve
      tags:
      - api
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemoTokens'
          description: ''
  /api/v1/auth/employer-login/:
    post:
      operationId: api_v1_auth_employer_login_create
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Login'
          description: ''
  /api/v1/auth/employer-register/:
    post:
      operationId: api_v1_auth_employer_register_create
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Register'
          description: ''
...
  /api/v1/auth/refresh/:
    post:
      operationId: api_v1_auth_refresh_create
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Refresh'
          description: ''
...
  /api/v1/auth/worker-login/:
    post:
      operationId: api_v1_auth_worker_login_create
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Login'
          description: ''
  /api/v1/auth/worker-register/:
    post:
      operationId: api_v1_auth_worker_register_create
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Register'
          description: ''
...
  /api/v1/workers/certificates/:
    post:
      operationId: api_v1_workers_certificates_create
      tags:
      - Workers
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CertificateUploadRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CertificateUploadRequest'
      security:
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
components:
  schemas:
...
    CertificateUploadRequest:
      type: object
      properties:
        certificates:
          type: array
          items:
            type: string
            format: binary
...
    DemoTokens:
      type: object
      properties:
        worker:
          type: object
          additionalProperties: {}
        employer:
          type: object
          additionalProperties: {}
        admin:
          type: object
          additionalProperties: {}
      required:
      - admin
      - employer
      - worker
...
    Login:
      type: object
      properties:
        identifier:
          type: string
      required:
      - identifier
    LoginRequest:
      type: object
      properties:
        identifier:
          type: string
          minLength: 1
        password:
          type: string
          writeOnly: true
          minLength: 1
      required:
      - identifier
      - password
...
    Refresh:
      type: object
      properties:
        refresh:
          type: string
      required:
      - refresh
    RefreshRequest:
      type: object
      properties:
        refresh:
          type: string
          minLength: 1
      required:
      - refresh
    Register:
      type: object
      properties:
        name:
          type: string
          maxLength: 120
        phone:
          type: string
        email:
          type: string
          format: email
      required:
      - name
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        phone:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          writeOnly: true
          minLength: 8
      required:
      - name
      - password
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT